requirements.txt
asgiref==3.2.10
Django==3.1.2
django-autoslug==1.9.8
django-ckeditor==6.0.0
django-crispy-forms==1.9.2
django-extensions==3.0.5
django-js-asset==1.2.2
django-taggit==1.3.0
django-widget-tweaks==1.4.8
mysqlclient @ file:///D:/Django/BlueBlogOverhauled/mysqlclient-1.4.6-cp38-cp38-win32.whl
Pillow==7.2.0
pytz==2020.1
sqlparse==0.3.1
